Social Network Data Analysis using Node XL and Tableau for Reliance Jio
Performed analytics on twitter and youtube network for data related to Reliance Jio a major telecom player in Indian Market to understand the disruptive effect of JIO. Initial analysis of the tweets before the start of Indian Premier League a cricket league in which JIOis a major sponsor which was earlier sponsored by Vodafone showed a negative trend. However, post the campaign it could be observed that the positive trend for JIO improved.