Predicting the number of runs required by Oakland A's in getting to baseball playoffs
Linear regression was performed to predict the runs scored, runs allowed and the number of wins required by a baseball team to get to the playoffs with the help of money ball dataset utilized by Paul De Podesta a pioneer in sabermetrics. The runs scored by a team was predicted using the baseball metrics of the team like on-base percentage, slugging rate. The runs allowed were predicted using the opponent team's on-base percentage and the opponent team's slugging rate. Number of wins acquired by a team was predicted using the run difference it., the difference between the runs scored and runs allowed. A scatter plot was used to understand the minimum run-difference required by a team to clear the play-off's.